SlugLighting
* Added Missing Slug Lighting to all slugs and simplified implementation of slugs
